major commit; bug fixes; stable versions;

+++ CRUDUserJSF/src/main/java/org/andvicoso/cruduser/model/dao/GenericDaoJPA.java
@@ -10,18 +10,24 @@ import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 import javax.persistence.TypedQuery;
 
-public abstract class GenericDaoJPA<O> implements GenericDao<O>, Serializable{
+public class GenericDaoJPA<O> implements GenericDao<O>, Serializable {
+	private static final String PERSISTENCE_UNIT = "cruduser";
+	private static EntityManagerFactory factory;
 	protected EntityManager entityManager;
 	private Class<O> type;
+	private String list;
 
 	protected GenericDaoJPA() {
-		EntityManagerFactory factory = Persistence
-				.createEntityManagerFactory("cruduser");
+		if (factory == null) {
+			factory = Persistence.createEntityManagerFactory(PERSISTENCE_UNIT);
+		}
 		entityManager = factory.createEntityManager();
-		// entityManager = new EntityManagerProducer().getEntityManager();
+
 		Type t = getClass().getGenericSuperclass();
 		ParameterizedType pt = (ParameterizedType) t;
+
 		type = (Class<O>) pt.getActualTypeArguments()[0];
+		list = "SELECT o FROM " + type.getSimpleName() + " o";
 	}
 
 	public void create(O obj) {
@@ -35,9 +41,7 @@ public abstract class GenericDaoJPA<O> implements GenericDao<O>, Serializable{
 	}
 
 	public List<O> list() {
-		String entityName = type.getSimpleName();
-		String q = "Select o from " + entityName + " o";
-		return createQuery(q).getResultList();
+		return createQuery(list).getResultList();
 	}
 
 	public O find(long id) {
@@ -45,14 +49,18 @@ public abstract class GenericDaoJPA<O> implements GenericDao<O>, Serializable{
 	}
 
 	public void remove(long id) {
-		O u = find(id);
-		entityManager.getTransaction().begin();
-		entityManager.remove(u);
-		entityManager.getTransaction().commit();
+		O u = entityManager.getReference(type, id);
+		if (u != null) {
+			entityManager.getTransaction().begin();
+			entityManager.remove(u);
+			entityManager.getTransaction().commit();
+		}
 	}
 
 	public void update(O obj) {
-		save(obj);
+		entityManager.getTransaction().begin();
+		entityManager.merge(obj);
+		entityManager.getTransaction().commit();
 	}
 
 	protected TypedQuery<O> createQuery(String q) {
